feat(web): extract MCP stats helpers into mcp-stats-lib
